Having and RollUp clause
HAVING Clause
The HAVING clause is used to filter records that work on aggregated data. It’s similar to the WHERE clause, but WHERE cannot be used with aggregate functions like SUM, COUNT, AVG, etc. Instead, HAVING is used after the GROUP BY clause to filter groups based on a condition.

ROLLUP Clause
The ROLLUP clause is an extension of the GROUP BY clause. It allows you to create subtotals and grand totals in your result set. This is particularly useful for generating reports that require hierarchical data aggregation.