SQL Join - Many to Many
A many-to-many relationship in SQL is used to associate multiple records in one table with multiple records in another table. This type of relationship is common in scenarios where entities can have multiple associations with each other. To implement a many-to-many relationship, you need an intermediate table, often called a junction or associative table, to link the two tables.